combine database and user initialization into a single resource
this allows for a working cleanup path; before, the user was being
deleted before the database, which would fail if any objects had been
created by the user. now, the database will be dropped before the user
is removed, so there shouldn't be any dependent objects at the time at
which the user is dropped.