use Feral snapshot to test fixes to response publishing
CloudFormation expects a fairly specific JSON object be uploaded to S3
as a response when a custom resource completes its work. The previous
milestone of Feral was not serializing it in quite the correct format,
but this snapshot should fix that problem. Once we report success, Feral
will publish a new milestone (which we'll want to adopt ASAP).